+++ vitest.config.js
@@ -1,13 +1,7 @@
 import { defineConfig } from 'vitest/config'
-import string from 'vite-plugin-string'
 
 export default defineConfig({
-  plugins: [
-    string({
-      include: ['**/*.html', '**/*.css'],
-    })
-  ],
   test: {
-    environment: 'jsdom',
+    include: ['**/*.{test,spec}.{js,mjs,cjs,ts,mts,cts,jsx,tsx}'],
   },
 })